requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
